Moved Signin and User to their own directory
