+++ app/src/main/java/com/adsamcik/signalcollector/activities/FeedbackActivity.java
@@ -17,7 +17,7 @@ import android.widget.TextView;
 import com.adsamcik.signalcollector.R;
 import com.adsamcik.signalcollector.utility.Assist;
 import com.adsamcik.signalcollector.network.Network;
-import com.adsamcik.signalcollector.utility.Signin;
+import com.adsamcik.signalcollector.signin.Signin;
 import com.adsamcik.signalcollector.utility.SnackMaker;
 
 import java.io.IOException;
