+++ app/src/main/java/com/adsamcik/signalcollector/activities/FileSharingActivity.java
@@ -18,7 +18,7 @@ import android.widget.Toast;
 
 import com.adsamcik.signalcollector.R;
 import com.adsamcik.signalcollector.file.DataFile;
-import com.adsamcik.signalcollector.utility.Signin;
+import com.adsamcik.signalcollector.signin.Signin;
 import com.adsamcik.signalcollector.utility.BottomSheetMenu;
 import com.adsamcik.signalcollector.file.Compress;
 import com.adsamcik.signalcollector.file.DataStore;
