+++ app/src/main/java/com/adsamcik/signalcollector/fragments/FragmentSettings.java
@@ -44,7 +44,7 @@ import com.adsamcik.signalcollector.interfaces.INonNullValueCallback;
 import com.adsamcik.signalcollector.interfaces.IValueCallback;
 import com.adsamcik.signalcollector.interfaces.IVerify;
 import com.adsamcik.signalcollector.network.Prices;
-import com.adsamcik.signalcollector.network.User;
+import com.adsamcik.signalcollector.signin.User;
 import com.adsamcik.signalcollector.services.ActivityService;
 import com.adsamcik.signalcollector.services.TrackerService;
 import com.adsamcik.signalcollector.utility.Assist;
@@ -58,7 +58,7 @@ import com.adsamcik.signalcollector.utility.Preferences;
 import com.adsamcik.signalcollector.file.DataStore;
 import com.adsamcik.signalcollector.interfaces.ITabFragment;
 import com.adsamcik.signalcollector.R;
-import com.adsamcik.signalcollector.utility.Signin;
+import com.adsamcik.signalcollector.signin.Signin;
 import com.adsamcik.signalcollector.utility.SnackMaker;
 import com.google.android.gms.common.SignInButton;
 import com.google.firebase.analytics.FirebaseAnalytics;
