+++ app/src/main/java/com/adsamcik/signalcollector/services/InstanceIDListenerService.java
@@ -1,10 +1,7 @@
 package com.adsamcik.signalcollector.services;
 
-import android.content.pm.PackageManager;
-import android.support.v4.content.ContextCompat;
-
 import com.adsamcik.signalcollector.network.Network;
-import com.adsamcik.signalcollector.utility.Signin;
+import com.adsamcik.signalcollector.signin.Signin;
 import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.iid.FirebaseInstanceIdService;
